服务器1.0.2版本 1.增加了获取用户轨迹的接口 2.增加了日报每日 每周 每月统计功能 3.增加了修改用户密码 基本信息功能 4.增加了日志记录

+++ src/main/java/com/wj/nongtian/entity/Project.java
@@ -24,7 +24,7 @@ public class Project {
     /**
      * 项目进度
      */
-    private int process;
+    private float process;
 
     /**
      * 项目创建者Id
@@ -95,11 +95,11 @@ public class Project {
         this.info = info;
     }
 
-    public int getProcess() {
+    public float getProcess() {
         return process;
     }
 
-    public void setProcess(int process) {
+    public void setProcess(float process) {
         this.process = process;
     }
